+++ test/storage.ts
@@ -18,10 +18,7 @@ class StorageTest {
     public before () {
         should();
         use(chaiAsPromised);
-        this.fxn = new Function({
-            accessKey: process.env.ACCESS_KEY!,
-            url: process.env.API_URL
-        });
+        this.fxn = new Function({ accessKey: process.env.ACCESS_KEY, url: process.env.API_URL });
     }
 
     @mocha.test